Update ep4-social-proposal-creation-of-foundation-working-groups-and-working-group-rules.md
This file should be deleted from term-4 in its entirety.  

Further, this is ep0.4, which I appears in term-4 in error due to the current naming convention where each file is named  in order (ep4...) as opposed (ep0.4).  Moreover, this naming convention has resulted in "ep4..." being erroneously duplicated in term-4, as a result I have begun to propose the renaming of the files to match the ep in lieu of the current naming convention which enumerates the file names as the ep was proposed.